error fix and some small modifications
