+++ frontend/lib/components/error/flutter_specific_error.dart
@@ -2,16 +2,18 @@ import 'package:flutter/material.dart';
 
 class CustomErrorWidget extends StatelessWidget {
   final String errorMessage;
+  final String originalRoute;
 
-  CustomErrorWidget({required this.errorMessage});
-
+  const CustomErrorWidget(
+      {Key? key, required this.errorMessage, required this.originalRoute})
+      : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Center(
       child: Directionality(
         textDirection: TextDirection.ltr,
         child: Padding(
-          padding: const EdgeInsets.all(8.0),
+          padding: EdgeInsets.all(8.0),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
@@ -32,6 +34,14 @@ class CustomErrorWidget extends StatelessWidget {
                 style: TextStyle(
                     fontSize: 18.0, color: Color.fromARGB(255, 158, 158, 158)),
                 textAlign: TextAlign.center,
+              ),
+              SizedBox(height: 5.0),
+              ElevatedButton(
+                onPressed: () {
+                  Navigator.pop(context);
+                  Navigator.pushNamed(context, originalRoute);
+                },
+                child: const Text('Erneut versuchen'),
               ),
             ],
           ),
